requirements.txt
Flask~=3.0
gunicorn~=21.2
python-dotenv~=1.0

# Realtime stack (to be used when enabling WebSockets)
flask-socketio~=5.3
eventlet~=0.36

# HTTP/JSON
requests~=2.32

# Database
SQLAlchemy~=2.0
psycopg[binary]~=3.2


